Move CutoutCell and TypicalCutoutCell classes to their own package
