+++ indicator/src/main/java/com/fuzz/indicator/cell/CutoutCell.java
@@ -13,11 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.fuzz.indicator;
+package com.fuzz.indicator.cell;
 
 import android.support.annotation.NonNull;
 import android.view.View;
 
+import com.fuzz.indicator.CutoutCellGenerator;
+import com.fuzz.indicator.CutoutViewIndicator;
+import com.fuzz.indicator.OffSetters;
 import com.fuzz.indicator.proxy.IndicatorOffsetEvent;
 
 /**
