+++ indicator/src/main/java/com/fuzz/indicator/text/TextCellGenerator.java
@@ -25,7 +25,7 @@ import android.widget.TextView;
 
 import com.fuzz.indicator.CutoutViewIndicator;
 import com.fuzz.indicator.CutoutViewLayoutParams;
-import com.fuzz.indicator.CutoutCell;
+import com.fuzz.indicator.cell.CutoutCell;
 import com.fuzz.indicator.CutoutCellGenerator;
 
 /**
